dev-lang/ghc: Support new libraries in BUMP_LIBRARIES
Previously, only supported bumping existing libraries as opposed to
adding new ones. Note that using this will likely require patching
hadrian as well so that it knows to build the new libraries.

+++ dev-lang/ghc/ghc-9.6.5.ebuild
@@ -236,7 +236,10 @@ bump_lib() {
 
 	einfo "Bumping ${pn} up to ${pv}"
 
-	mv "${dir}"/"${pn}" "${WORKDIR}"/"${pn}".old || die
+	if [[ -d "${dir}"/"${pn}" ]] # if the library exists already, move it out of the way
+	then
+		mv "${dir}"/"${pn}" "${WORKDIR}"/"${pn}".old || die
+	fi
 	mv "${WORKDIR}"/"${p}" "${dir}"/"${pn}" || die
 }
